devin-demo/function/requirements.txt
google-cloud-storage>=2.0.0,<3.0.0
